+++ community/pmc-tool/main/templates/results.html
@@ -18,20 +18,18 @@
     {% endfor %}
     </ul>
 
-    <h1>Pull Request Authors (Last Week):</h1>
+    <h1>PR Authors in last week(all):</h1>
     <ul>
     {% for author, date in first_pr_authors %}
         <li>{{ author }}: {{ date }}</li>
     {% endfor %}
     </ul>
 
-    <h1>New Authors (First PR in Last Week):</h1>
+    <h1>New Authors(First PR):</h1>
     <ul>
     {% for author in new_authors %}
         <li>{{ author }}</li>
     {% endfor %}
     </ul>
-
-
 </body>
 </html>
